# src/main/resources/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Petme Api
  license:
    name: www.petme.com
servers:
  - url: http://localhost:8080
paths:
  /users:
    get:
      summary: 사용자 조회
      operationId: listUsers
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: 사용자 추가
      operationId: createUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        201:
          description: Created
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{userId}:
    get:
      summary: 사용자 조회
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: 사용자 삭제
      operationId: deleteUsers
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        204:
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        loginId:
          type: string
        userName:
          type: string
        password:
          type: string
        email:
          type: string
        phone:
          type: string
        residentRegistrationNumber:
          type: string
        streetNameAddress:
          type: string
        lotNumberAddress:
          type: string
        gender:
          type: string
        receiveInformationYn:
          type: string
        receiveEventYn:
          type: string
        accessToken:
          type: string
        oauthType:
          type: string
        petType:
          type: string
        currentPoint:
          type: integer
        lastLogin:
          type: string
        loginTimes:
          type: integer
        withdrawalDate:
          type: string
        withdrawalCause:
          type: string
        useYn:
          type: string
        modifyTime:
          type: string
        registerTime:
          type: string
      required:
        - loginId
        - currentPoint
        - loginTimes
        - modifyTime
        - registerTime
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User"